Do not follow tags on default push
Turns out, if the local branch is out of date with the upstream branch,
`--follow-tags` can succeed for the tag push, but fail for the commit
push. This leaves you in the incredibly shitty position of having pushed
a tag that points to a commit that doesn't exist on the branch you
wanted it to exist on.

This can be solved by using `--atomic`, but unfortunately there is no
`push.atomic` option in `.gitconfig` currently.