Completions seem to be working as intended
That is to say, brew completions don't work because of
https://github.com/zhaofengli/nix-homebrew/issues/77 and mas completions
don't work because the Homebrew package only provides bash completions.

I thought zsh could use the bash completions automatically, but it seems
that's not the case. The mas completions use functions provided by the
bash-completions package, and that's a whole rabbit hole.